requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Login>=0.6.3
Flask-WTF>=1.2.0
WTForms>=3.1.0
python-dotenv>=1.0.0
Flask-CORS>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
cryptography>=41.0.0
requests>=2.31.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0